requirements.txt
resemblyzer==0.1.4
nats-py==2.9.0
numpy==1.26.4
scipy==1.13.1
soundfile==0.12.1
httpx==0.27.0
